Moved phone number validation to its own file
Moved phone number validation to its own file/module to split code imports better. Phone number validation uses `google-libphonenumber` imported package which is very large.
